requirements.txt
fastapi==0.128.0
pydantic==2.12.5
uvicorn==0.40.0
llama-cpp-python
